Added hyperlink type, ScrollerBar
