+++ packages/theme/styles/_layouts.scss
@@ -518,6 +518,7 @@ input.search {
 .max-h-60 { max-height: 15rem; }
 .max-w-30 { max-width: 7.5rem; }
 .max-w-60 { max-width: 15rem; }
+.max-w-80 { max-width: 20rem; }
 .max-w-240 { max-width: 60rem; }
 .clear-mins {
   min-width: 0;
